feat(rust): update cli arguments for more realistic usage
