+++ implementations/rust/daemon/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+url = "2.1.1"
 structopt = { version = "0.3.17", default-features = false }
-# ockam-message = { path = "../message", version = "0.1.0" }
+ockam-message = { path = "../message", version = "0.1.0" }
 # ockam-vault = { path = "../vault", version = "0.1.0" }
