+++ implementations/rust/daemon/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod args;
+pub mod cli;
